move plan switching hooks into trigger class
